requirements.txt
torch
torchvision
segmentation_models_pytorch
albumentations
gradio
huggingface_hub
Pillow>=9.0,<11.0
numpy
